backend/permits/requirements.txt
# AI-Powered Permit Extraction Dependencies

# OpenAI API
openai>=1.0.0

# Image Processing
Pillow>=10.0.0

# PDF Processing
PyPDF2>=3.0.0
pdf2image>=1.16.0

# Note: pdf2image requires poppler-utils to be installed on the system:
# Ubuntu/Debian: sudo apt-get install poppler-utils
# macOS: brew install poppler
# Windows: Download from https://github.com/oschwartz10612/poppler-windows/releases/
